Dirtied up the code a bit trying to fix a bug.
At some point in the future, need to go back and use a temporary object to store query results. I won't bother with this until it actually causes a problem.

Biggest change is that I moved the Rocker C++ stuff to Rockerxx so it wouldn't keep loading the wrong .so file.